added input class
