+++ PixelPhysics/bin/.gitignore
@@ -1,1 +1,2 @@
 /main/
+/workers/
