+++ PixelPhysics/src/main/GamePanel.java
@@ -32,7 +32,7 @@ import workers.GraphicsWorker;
 import workers.PackingWorker;
 import workers.PullPhysicsWorker;
 
-public class GamePanel extends JPanel implements MouseListener,KeyListener,ActionListener{
+public class GamePanel extends JPanel{
 
 	/**
 	 * 
@@ -41,27 +41,17 @@ public class GamePanel extends JPanel implements MouseListener,KeyListener,Actio
 	private ExecutorService executorGraphics = Executors.newCachedThreadPool();
 	private ExecutorService executorBlobs = Executors.newCachedThreadPool();
 	private static final long serialVersionUID = 1L;
-	boolean lmbHeld = false;
-	boolean rmbHeld = false;
-	public final int maxTimer = 30;
 	public int[] cooldowns = new int[10];
 	public ArrayList<Particle> particleArray = new ArrayList<Particle>();
-	public double lastLag1 = 0;
-	public double lastLag2 = 0;
-	public double lastLag3 = 0;
-	public int RGB_switch = 1;
 	public boolean flag = false;
-	public Point lastClick = null;
 	public ArrayList<Point2D> pullQueue = new ArrayList<Point2D>();
 	public ArrayList<Point2D> pushQueue = new ArrayList<Point2D>();
-	public ArrayList<Wall> wallArray = new ArrayList<Wall>();
 	public static final Random rand = new Random();
 	public BitSet keySet = new BitSet(256);
 	int[][] RGBs = new int[1920][1080];
 	DecimalFormat df2 = new DecimalFormat("0.00");
 	DecimalFormat df3 = new DecimalFormat("0.000");
 	BufferedImage paintBuffer;
-	public boolean showStats = true;
 	public GamePanel(int w,int h,int m,int s){
 		Properties.cores = Runtime.getRuntime().availableProcessors();
 		Properties.maxPixels = m;
@@ -73,37 +63,12 @@ public class GamePanel extends JPanel implements MouseListener,KeyListener,Actio
 		setPreferredSize(d);
 		this.setFocusable(true);
 	}
-	public void switchShift(){
-		RGB_switch = rand.nextInt(3);
-		Properties.shiftAmount = -1;
-		if(getPositive(RGB_switch)){
-			Properties.shiftAmount = 1;
-		}
-	}
-	public boolean getPositive(int index){
-		if(Properties.RGB[index] < 127){
-			return true;
-		}
-		return false;
-	}
-	public void shiftColor(){
-		Properties.RGB[RGB_switch] += Properties.shiftAmount;
-		if(Properties.RGB[RGB_switch] > 255){
-			Properties.RGB[RGB_switch] = 255;
-			switchShift();
-		}
-		else if(Properties.RGB[RGB_switch] < 1){
-			Properties.RGB[RGB_switch] = 1;
-			switchShift();
-		}
-		else if(rand.nextInt(100) < 1){
-			switchShift();
-		}
-	}
+	
 	public void init(){
-		addMouseListener(this);
+		Input in = new Input();
+		addMouseListener(in);
 		int a = Properties.maxPixels;
-		addKeyListener(this);
+		addKeyListener(in);
 		setBackground(Color.BLACK);
 		for(int i = 0; i < a;i++){
 			spawnify();
@@ -123,34 +88,24 @@ public class GamePanel extends JPanel implements MouseListener,KeyListener,Actio
 		else{
 			drawParticles(g);
 		}
-		for(int i = 0; i < wallArray.size();i++){
-			Wall w = wallArray.get(i);
-			g.setColor(w.c);
-			g.fillPolygon(w.p);
-		}
 		long t1 = System.nanoTime();
-		lastLag1 = (t1 - t0) / 1000000D;
+		double lastLag1 = (t1 - t0) / 1000000D;
 		double fps = (16D / (double)lastLag1) * 60;
 		g.setColor(Color.WHITE);
 		g.drawString(df2.format(lastLag1) + " (" +(int)fps+")",0, 20);
-		if(showStats){
+		if(Properties.showStats){
 			g.drawString(Properties.RGB[0] + " " + Properties.RGB[1] + " " + Properties.RGB[2] + " " + Properties.glowStrength, getWidth() / 2, (getHeight() / 2));
 			g.drawString(df3.format(Properties.frictionStrength) + " - " + df2.format(Properties.pullStrength) + " - " + Properties.size + " - " + df2.format(Properties.timeSpeed), getWidth() / 2, (getHeight() / 2) + 20);	
 		}
 	}
-	public void addWall(int x1,int y1, int x2, int y2){
-		Wall w = new Wall(x1,y1,x2,y2);
-		wallArray.add(w);
-	}
 	public void update(){
-		Properties.glowRadius = (Properties.size * 2);
 		this.updateParticles();
-		this.updateThemkeys();
+		Input.updateThemkeys();
 		this.doMouse();
 		this.repaint();
 		flag = !flag;
 		if(flag){
-			shiftColor();
+			Properties.shiftColor();
 		}
 	}
 	public void drawBlobsWithWorkers(Graphics gg){
@@ -198,19 +153,25 @@ public class GamePanel extends JPanel implements MouseListener,KeyListener,Actio
 		gg.drawImage(b3, 0, h/2, this);
 		gg.drawImage(b4, w/2, h/2, this);
 		long t2 = System.nanoTime();
-		lastLag2 = (t1 - t0) / 1000000D;
-		lastLag3 = (t2 - t1) / 1000000D;
-
 	}
-	public void drawParticles(Graphics g){
+	public void drawParticles(Graphics gg){
 		if(paintBuffer != null){
 			paintBuffer = null;
 		}
 		ArrayList<Particle> pA = packify(particleArray);
-		g.setColor(new Color(Properties.RGB[0],Properties.RGB[1],Properties.RGB[2]));
+		gg.setColor(new Color(Properties.RGB[0],Properties.RGB[1],Properties.RGB[2]));
+		int w = getWidth();
+		int h = getHeight();
 		for(int i = 0; i < pA.size();i++){
 			Particle p = pA.get(i);
-			g.fillRect((int)p.x ,(int)p.y, Properties.size,Properties.size);
+			if(Properties.rainbow){
+//				g.setColor(p.color);
+				int r = 0;
+				int g = 0;
+				int b = 0;
+				gg.setColor(new Color(r,g,b));
+			}
+			gg.fillRect((int)p.x ,(int)p.y, Properties.size,Properties.size);
 		}
 	}
 	public void drawParticlesPaint(Graphics g){
@@ -219,55 +180,15 @@ public class GamePanel extends JPanel implements MouseListener,KeyListener,Actio
 			paintBuffer = new BufferedImage(1920, 1080, BufferedImage.TYPE_INT_RGB);
 		}
 		Graphics gg = paintBuffer.getGraphics();
+		gg.setColor(new Color(Properties.RGB[0],Properties.RGB[1],Properties.RGB[2]));
 		for(int i = 0; i < pA.size();i++){
 			Particle p = pA.get(i);
-			gg.setColor(new Color(Properties.RGB[0],Properties.RGB[1],Properties.RGB[2]));
+			if(Properties.rainbow){
+				gg.setColor(p.color);
+			}
 			gg.fillRect((int)p.x, (int)p.y, Properties.size, Properties.size);
 		}	
 		g.drawImage(paintBuffer, 0, 0, null);
-	}
-	public BufferedImage getImage(ArrayList<Particle> pA){
-		BufferedImage buffImage = new BufferedImage(getWidth(),getHeight(),BufferedImage.TYPE_INT_RGB);
-		Graphics gg = buffImage.getGraphics();		
-		gg.setColor(Color.BLUE);
-		for(int i = 0; i < pA.size();i++){
-			Particle p  = pA.get(i);
-			//			gg.setColor(p.color);
-			gg.fillRect((int)p.x ,(getHeight() - (int)p.y), Properties.size,Properties.size);
-		}
-		return buffImage;
-
-	}
-	public BufferedImage[] getImageWithWorker(ArrayList<Particle> pA){
-		BufferedImage[] buffImages = new BufferedImage[4];
-		int q = pA.size() / 4;
-		int h = pA.size() / 2;
-		int f = pA.size();
-		int width = getWidth();
-		int height = getHeight();
-		int r = Properties.RGB[0];
-		int g = Properties.RGB[1];
-		int b = Properties.RGB[2];
-		Future<BufferedImage> w1 = executorGraphics.submit(new GraphicsWorker((new ArrayList<Particle>(pA.subList(0, h))),width,height,Properties.size,r,g,b));
-		Future<BufferedImage> w2 = executorGraphics.submit(new GraphicsWorker((new ArrayList<Particle>(pA.subList(h, f))),width,height,Properties.size,r,g,b));
-		Future<BufferedImage> w3 = executorGraphics.submit(new GraphicsWorker((new ArrayList<Particle>(pA.subList(0, h))),width,height,Properties.size,r,g,b));
-		Future<BufferedImage> w4 = executorGraphics.submit(new GraphicsWorker((new ArrayList<Particle>(pA.subList(0, h))),width,height,Properties.size,r,g,b));
-
-
-		do{
-
-		}while(!w1.isDone() && !w2.isDone() && !w3.isDone() && !w4.isDone());
-
-		try {
-			buffImages[0] = w1.get();
-			buffImages[1] = w2.get();
-			buffImages[2] = w3.get();
-			buffImages[3] = w4.get();
-		} catch (InterruptedException | ExecutionException e) {
-			e.printStackTrace();
-		}
-		return buffImages;
-
 	}
 	//TODO When multiple particle are in the same spot it swaps between which one it draws, and when different colors = flickering :(
 	//TODO With packing, particles are non-existant? wut
@@ -322,9 +243,6 @@ public class GamePanel extends JPanel implements MouseListener,KeyListener,Actio
 		}
 		return p1;
 	}
-	public ArrayList<Particle> getParticles(){
-		return particleArray;
-	}
 	public void updateParticles(){
 
 		for(int i = 0; i < pullQueue.size();i++){
@@ -352,20 +270,6 @@ public class GamePanel extends JPanel implements MouseListener,KeyListener,Actio
 		}
 
 	}
-	//	public void pull(double x, double y, double mult){
-	//		double angle = 1;	
-	//		double deltaX;
-	//		double deltaY;
-	//		mult *= timeSpeed * pullStrength;
-	//		for(int i = 0; i < particleArray.size();i++){
-	//			Particle p = particleArray.get(i);
-	//			angle = FastMath.atan2((float)(p.y - y), (float)(p.x - x));
-	//			deltaX = net.jafama.FastMath.cos(angle);
-	//			deltaY = net.jafama.FastMath.sin(angle);
-	//			p.speedX -= deltaX * mult;							
-	//			p.speedY -= deltaY * mult;      					 
-	//		}
-	//	}
 	public void pullWithWorkers(double x, double y, double mult){
 		int q = particleArray.size() / 4;
 		int h = particleArray.size() / 2;
@@ -418,205 +322,21 @@ public class GamePanel extends JPanel implements MouseListener,KeyListener,Actio
 		}
 		return false;
 	}
-	public void updateThemkeys(){
-		for(int i = 0 ; i < cooldowns.length; i++){
-			cooldowns[i]--;
-			if(cooldowns[i] < 0){
-				cooldowns[i] = 0;
-			}
-		}
-		if(keySet.get(KeyEvent.VK_UP)){
-			Properties.timeSpeed += 0.01;
-		}
-		if(keySet.get(KeyEvent.VK_DOWN)){
-			Properties.timeSpeed -= 0.01;
-			if(Properties.timeSpeed < 0.01){
-				Properties.timeSpeed = 0.01;
-			}
-		}
-		if(keySet.get(KeyEvent.VK_RIGHT)){
-			Properties.timeSpeed = 0.5;
-		}
-		if(keySet.get(KeyEvent.VK_LEFT)){
-			Properties.timeSpeed = 0;
-		}
-		if(keySet.get(KeyEvent.VK_W)){
-			if(Math.abs(Properties.pullStrength) - 1 <= 0){
-				Properties.pullStrength += 0.003;
-			}
-			else{
-				Properties.pullStrength += Math.abs(Properties.pullStrength) / 100D;
-			}
-			if(Properties.pullStrength > Properties.pulls[2]){
-				Properties.pullStrength = Properties.pulls[2];
-			}
-		}
-		if(keySet.get(KeyEvent.VK_R)){
-			Properties.pullStrength = Properties.pulls[1];
-		}
-		if(keySet.get(KeyEvent.VK_F)){
-			Properties.frictionStrength = Properties.frictions[1];
-		}
-
-		if(keySet.get(KeyEvent.VK_S)){
-			if(Math.abs(Properties.pullStrength) - 1 <= 0){
-				Properties.pullStrength -= 0.003;
-			}
-			else{
-				Properties.pullStrength -= Math.abs(Properties.pullStrength) / 100D;
-			}
-			if(Properties.pullStrength < Properties.pulls[0]){
-				Properties.pullStrength = Properties.pulls[0];
-			}
-
-		}
-		if(keySet.get(KeyEvent.VK_A)){
-			Properties.frictionStrength -= Properties.frictionStrength / 50D;
-			if(Properties.frictionStrength < Properties.frictions[0]){
-				Properties.frictionStrength = Properties.frictions[0];
-			}
-		}
-		if(keySet.get(KeyEvent.VK_D)){
-			Properties.frictionStrength += Properties.frictionStrength / 50D;
-			if(Properties.frictionStrength > Properties.frictions[2]){
-				Properties.frictionStrength = Properties.frictions[2];
-			}
-		}
-		if(keySet.get(KeyEvent.VK_COMMA)){
-			if(cooldowns[0] == 0){
-				cooldowns[0] = maxTimer;
-				Properties.size--;
-				if(Properties.size < 1){
-					Properties.size = 1;
-				}
-			}
-		}
-		if(keySet.get(KeyEvent.VK_PERIOD)){
-			if(cooldowns[1] == 0){
-				cooldowns[1] = maxTimer;
-				Properties.size++;
-				if(Properties.size > Properties.maxSize){
-					Properties.size = Properties.maxSize;
-				}
-			}
-		}
-		if(keySet.get(KeyEvent.VK_G)){
-			if(cooldowns[2] == 0){
-				cooldowns[2] = maxTimer;
-				Properties.glow = !Properties.glow;
-			}
-		}
-		if(keySet.get(KeyEvent.VK_J)){
-			if(cooldowns[3] == 0){
-				cooldowns[3] = maxTimer;
-				Properties.compound = !Properties.compound;
-			}
-		}
-		if(keySet.get(KeyEvent.VK_L)){
-			if(cooldowns[4] == 0){
-				cooldowns[4] = maxTimer / 8;
-				Properties.glowStrength -= 2;
-				if(Properties.glowStrength < 2){
-					Properties.glowStrength = 2;
-				}
-			}
-		}
-		if(keySet.get(KeyEvent.VK_O)){
-			if(cooldowns[5] == 0){
-				cooldowns[5] = maxTimer / 8;
-				Properties.glowStrength += 2;
-				if(Properties.glowStrength > 255){
-					Properties.glowStrength = 255;
-				}
-			}
-		}
-		if(keySet.get(KeyEvent.VK_U)){
-			if(cooldowns[6] == 0){
-				cooldowns[6] = maxTimer;
-				Properties.pixelized= !Properties.pixelized;
-			}
-		}
-		if(keySet.get(KeyEvent.VK_Y)){
-			if(cooldowns[7] == 0){
-				cooldowns[7] = maxTimer;
-				Properties.LSD = !Properties.LSD;
-			}
-		}
-		if(keySet.get(KeyEvent.VK_P)){
-			if(cooldowns[8] == 0){
-				cooldowns[8] = maxTimer;
-				Properties.paint = !Properties.paint;
-			}
-		}
-		if(keySet.get(KeyEvent.VK_E)){
-			shiftColor();
-
-		}
-		if(keySet.get(KeyEvent.VK_H)){
-			Properties.RGB[0] = 255;
-			Properties.RGB[1] = 70;
-			Properties.RGB[2] = 0;
-		}
-		if(keySet.get(KeyEvent.VK_Z)){
-			if(cooldowns[9] == 0){
-				cooldowns[9] = maxTimer;
-				showStats = !showStats;
-			}
-		}
-	}
 	public void doMouse(){
-		if(lmbHeld){
+		if(Properties.lmbHeld){
 			Point p = MouseInfo.getPointerInfo().getLocation() ;
 			int x = (int) (p.getX() - this.getLocationOnScreen().getX());
 			int y = (int) (p.getY() - this.getLocationOnScreen().getY());
 			pullQueue.add(new Point(x,y));
 		}
-		else if(rmbHeld){
+		else if(Properties.rmbHeld){
 			Point p = MouseInfo.getPointerInfo().getLocation() ;
 			int x = (int) (p.getX() - this.getLocationOnScreen().getX());
 			int y = (int) ((int) p.getY() + this.getLocationOnScreen().getY()) - 57;
 			pushQueue.add(new Point(x,y));
 		}
 	}
-	@Override
-	public void mousePressed(MouseEvent e) {
-		if(e.getButton() == MouseEvent.BUTTON1){
-			lmbHeld = true;
-		}
-		if(e.getButton() == MouseEvent.BUTTON3){
-			rmbHeld = true;
-		}
-		if(e.getButton() == MouseEvent.BUTTON2){
-			System.out.println(lastClick);
-			Point p = e.getPoint();
-			if(lastClick == null){
-				lastClick = p;
-			}
-			else{
-				addWall(lastClick.x, lastClick.y, p.x, p.y);
-				lastClick = null;
-			}
-		}
-	}
 
-	@Override
-	public void keyPressed(KeyEvent e) {
-		keySet.set(e.getKeyCode(),true);
-	}
-	@Override
-	public void keyReleased(KeyEvent e) {
-		keySet.set(e.getKeyCode(),false);
-	}
-	@Override
-	public void mouseReleased(MouseEvent e) {
-		lmbHeld = false;
-		rmbHeld = false;
-	}
-	@Override
-	public void mouseExited(MouseEvent e) {
-		lmbHeld = false;
-		rmbHeld = false;
-	}
 	public static Color randomColor(){
 		Random rand = new Random();
 		int[] rgb = new int[3];
@@ -642,18 +362,5 @@ public class GamePanel extends JPanel implements MouseListener,KeyListener,Actio
 		return Math.sqrt((x1-x2)*(x1-x2) + (y1-y2)*(y1-y2));
 	}
 
-	@Override
-	public void mouseEntered(MouseEvent e) {}
-	@Override
-	public void mouseClicked(MouseEvent e) {}
-	@Override
-	public void keyTyped(KeyEvent e) {}
-	@Override
-	public void actionPerformed(ActionEvent e) {
-	}
-	public static <T extends Enum<?>> T randomEnum(Class<T> clazz){
-		int x = rand.nextInt(clazz.getEnumConstants().length);
-		return clazz.getEnumConstants()[x];
-	}
 
 }
